Fix block-vs-point evaluation mismatches and overflow bugs flagged by review
Addresses 10 unresolved Copilot review comments on PR #467 that prior review
passes on this branch had not picked up (they postdated the last review's
memory checkpoint and were never re-checked against github_pr_review_comments).

Block-evaluation (computeValues) vs point-evaluation (computeValue) agreement:
- Minus, Difference: narrow to 32 bits when getType() == Integer.class instead
  of always negating/subtracting as long, reproducing Java's int wraparound
  (truncation commutes with +/-/negation, so narrowing once at the end matches
  narrowing at every step).
- And: narrow through (long) before (int) instead of casting double straight
  to int, matching Number#intValue()'s low-32-bit truncation instead of
  saturating.

Exact-value handling for kernel series detection:
- KernelSeriesMatcher: the point-evaluation fallback (triggered when block
  evaluation hits an inexact intermediate) now goes through the same
  exact-double check block evaluation uses, refusing to match rather than
  aliasing distinct large longs into a false pattern.
- IndexSequence.getExpression: same fix for the array-sequence-based
  recognition path.
- KernelSeriesCache.referenceSeries: refuses to cache an int sequence whose
  values can't be represented exactly as double, since the lookup table is
  double-backed regardless of isInt.
- Added IndexRange.exact(Number) as the shared conversion used by all three.

Overflow/edge-case fixes:
- ArithmeticIndexSequence.dividedExactly: guard against Long.MIN_VALUE, whose
  negation is a no-op and previously caused infinite recursion.
- IndexRange.partition: capacity calculation and loop counter no longer
  overflow int near Integer.MAX_VALUE.
- ArithmeticGenerator.divide, and the equivalent Quotient.create fold over a
  Mod numerator: a negative divisor no longer produces an invalid negative
  ArithmeticGenerator granularity.
- Quotient.create's product/divisor fold: the constant-factor multiplication
  is now overflow-checked via Math.multiplyExact rather than a plain product,
  after determining that gating the fold on the numerator's own bound (as
  originally suggested) regressed the branch's own mixedRadixFolds test,
  since a bare kernel index has no derivable bound in either the safe or
  unsafe case.
- Product.upperBound and the lowerBound fast path: also overflow-checked via
  Math.multiplyExact, since the Quotient fix above depends on these being
  trustworthy.

Added regression tests for each fix in ArithmeticSequenceDerivationTests,
IndexRangeEvaluationTests, IndexRangeTest, KernelSeriesMatcherTests, and
KernelSeriesCacheTest. Replied to each Copilot comment with what changed.

+++ base/code/src/main/java/io/almostrealism/expression/And.java
@@ -193,6 +193,15 @@ public class And extends BinaryExpression<Integer> {
 		return children[0].intValue() & children[1].intValue();
 	}
 
+	/**
+	 * {@inheritDoc}
+	 *
+	 * <p>Narrows each operand through {@code long} before narrowing to {@code int}, matching
+	 * {@link Number#intValue()}'s low-32-bit truncation. Narrowing the {@code double} straight
+	 * to {@code int} would instead saturate to {@code Integer.MAX_VALUE}/{@code MIN_VALUE} for
+	 * an operand whose value exceeds the {@code int} range, disagreeing with
+	 * {@link #computeValue(IndexValues)}.</p>
+	 */
 	@Override
 	protected double[] computeValues(IndexRange range) {
 		double[] left = getChildren().get(0).values(range);
@@ -200,7 +209,7 @@ public class And extends BinaryExpression<Integer> {
 		double[] out = new double[range.getLength()];
 
 		for (int i = 0; i < out.length; i++) {
-			out[i] = ((int) left[i]) & ((int) right[i]);
+			out[i] = ((int) (long) left[i]) & ((int) (long) right[i]);
 		}
 
 		return out;
